refactor: Reintegration of the header section
