+++ src/App.tsx
@@ -5,11 +5,13 @@ import { Footer } from "./components/Footer/Footer"
 import { Navbar } from "./components/Navbar/Navbar"
 import { Projects } from "./components/Projects/Projects"
 import { ThemeProvider } from "./context/ThemeProvider"
+import { Header } from './components/Header/Header';
 
 export const App = () => (
   <ThemeProvider>
     <div id="home" className="back-gradient"></div>
     <Navbar/>
+    <Header/>
     {/* <AboutMe/> */}
     {/* <Experience/> */}
     {/* <Projects/> */}
