--- src/components/ContactButton/ContactButtonStyles.css
@@ -1,21 +1,0 @@
-.contact-btn {
-  width: 11rem;
-  padding: 1rem 1.25rem;
-  letter-spacing: .08rem;
-  text-align: center;
-  text-decoration: none;
-  font-family: var(--cheddar-font);
-  font-size: var(--hero-text-size);
-  line-height: var(--hero-text-size);
-  font-weight: var(--cheddar-regular);
-  border-radius: 4px;
-  background-color: var(--rock-red);
-  color: var(--main-white);
-  cursor: pointer;
-  transition: .15s ease-in-out;
-
-  &:hover {
-    background-color: #cc0000;
-    color: var(--main-white);
-  }
-}
