+++ src/index.css
@@ -35,7 +35,7 @@
   --outfit-medium: 500;
   --outfit-semi-bold: 600;
   --outfit-bold: 700;
-  --title-size: 3rem;
+  --title-size: 2.75rem;
   --subtitle-size: 2rem;
   --hero-text-size: 1.5rem;
   --about-me-size: 1.1rem;
